requirement.txt
numpy
pandas
tqdm
lmdb
opencv-python
pillow
wandb
easydict
scipy
pycocotools
imageio
einops
rotary_embedding_torch
